mqttlogic: change loop prevention
This commit changes the endless loop prevention that could occur when
an item keeps changing itself due to its own change ...
The old mechanism would prevent the logic to run when the item itself
changed, but this may be necessary to start/trigger timers etc.
The new mechanism will run the logic, and warn when the output would
change.