rpnlogic: refactor stack, keep string representation in parallel
This commit changes the handling of strings. Wherever feasible, the string
representation is kept around (as const char *).
This allows much some operators (like ==) to operate on strings
when both operands have a string around.
Due to the refactoring to keep double & string around, helper functions
are implemented to manipulate the stack, and stack underflows are checked
in the helper functions now.